feat: initial implementation - logging dependencies for files
First pass logs the count of dependencies overall, the unique set of files visited, and all of the dependency paths encountered. As a start this exposes enough information to determine quickly-enough whether or not the dependency tree for a particular file might be greater than you expect and should you what they are.